// src/types/bindings.ts
export type AnimeListInfo = {
	image_url: string,
	judul: string,
	anime_name: string,
	studio: string,
	score: string,
	musim_rilis: string,
	genres: GenreInformation[],
	sinopsis: string[],
};
// ...
export type GenreInformation = {
	name: string,
	display: string,
};
// ...
export type Season = "Winter" | "Summer" | "Fall" | "Spring";